DJANGO User Activity Logs/requirements.txt
Django==5.0.4
djangorestframework==3.15.1

# PostgreSQL driver
psycopg2-binary==2.9.9

# Redis cache support
django-redis==5.4.0

# Environment variable support
python-dotenv==1.0.1

# Testing tools
pytest==8.2.0
pytest-django==4.8.0
